Had to adjust classes for api call
