+++ app/src/main/java/com/Lambda/rv_camping/model/Properties.kt
@@ -1,6 +1,6 @@
 package com.Lambda.rv_camping.model
 
-data class Properties (
+data class Property (
     val id: Int,
     var property_name: String,
     var description: String,
@@ -11,4 +11,8 @@ data class Properties (
     var price: Int,
     var rating: Number,
     var owner_id: Int
+)
+
+data class Properties(
+    val properties: List<Property>
 )
