+++ app/src/main/java/com/Lambda/rv_camping/ui/controllers/LoginController.kt
@@ -5,7 +5,10 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
+import androidx.recyclerview.widget.LinearLayoutManager
 import com.Lambda.rv_camping.R
+import com.Lambda.rv_camping.adapter.PropertiesAdapter
+import com.Lambda.rv_camping.model.Properties
 import com.Lambda.rv_camping.model.User
 import com.Lambda.rv_camping.model.UserLogin
 import com.Lambda.rv_camping.model.UserResponse
@@ -19,6 +22,7 @@ import com.bluelinelabs.conductor.ControllerChangeHandler
 import com.bluelinelabs.conductor.ControllerChangeType
 import com.bluelinelabs.conductor.RouterTransaction
 import com.bluelinelabs.conductor.changehandler.HorizontalChangeHandler
+import kotlinx.android.synthetic.main.activity_main.view.*
 import kotlinx.android.synthetic.main.controller_login.view.*
 import retrofit2.Call
 import retrofit2.Callback
@@ -40,6 +44,10 @@ class LoginController : Controller(){
     lateinit var password: String
     
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup): View {
+
+
+        getAllProperties()
+
         val view = inflater.inflate(R.layout.controller_login, container, false)
         view.pb_login.gone()
 
@@ -64,6 +72,8 @@ class LoginController : Controller(){
 
         //TODO REMOVE THIS ON OFFICIAL RELEASE!!!
         view.btn_login_skip.setOnClickListener {
+
+            getAllProperties()
             router.pushController(RouterTransaction.with(MainController())
                 .pushChangeHandler(HorizontalChangeHandler())
                 .popChangeHandler(HorizontalChangeHandler()))
@@ -162,6 +172,35 @@ class LoginController : Controller(){
                 }
             }
 
+        })
+    }
+
+    fun getAllProperties(){
+        val call: Call<Properties> = ApiBuilder.create().getAllProperties()
+        call.enqueue(object: Callback<Properties>{
+            override fun onFailure(call: Call<Properties>, t: Throwable) {
+                Log.i("Properties", "OnFailure ${t.message}")
+            }
+
+            override fun onResponse(
+                call: Call<Properties>,
+                response: Response<Properties>
+            ) {
+                if(response.isSuccessful){
+                    Log.i("Properties", "onResponseSuccessful ${response.body()}")
+                    val properties: List<Properties>?
+//                    properties?.forEach {
+//                        MainController.propertyList.add(
+////                            Properties(it.id, it.property_name, it.description, it.address,
+////                                it.city, it.state, it.image, it.price, it.rating, it.owner_id)
+//                        )
+                    }
+
+                    view?.vRecycle?.apply {
+                        layoutManager = LinearLayoutManager(activity)
+                        //adapter = PropertiesAdapter(MainController.propertyList)
+                    }
+                }
         })
     }
